requirements.txt
fastapi
uvicorn
numpy
pandas
psycopg2-binary
scikit-learn
google-cloud-storage
google-generativeai
python-dotenv
pydantic
pydantic-settings
faiss-cpu